lf-assoc:  Improve performance, don't change mac for created sta by default
There is no need to map sta name to port-id, and that method was quite
costly since it requested IDs for all ports.  Instead, just use the
port-name, which is going to work more reliably anyway.

And, don't change MAC of already-created station unless user specifically
requests this with --change_mac