firemod:  Attempt to fix INC path issue.
On my system that does not have the directory /home/lanforge/scripts, the
lf_firemod.pl would fail saying it cannot find Net::Telnet, even though it
is installed.

The problem turns out to be that if you have "use lib foo" and foo does not exist,
or is not readable, then perl craps out immediately instead of trying other
places in @INC.  I think this 'use if' logic fixes that issue based on
internet reading.  It fixes the issue on my system, but I'm not certain
it works properly on systems that do depend on /home/lanforge/scripts/
existing, so that will need some more testing.

Signed-off-by: Ben Greear <greearb@candelatech.com>