cv-scripts:  create tests with 'do not load old cfg' option.
This puts test in known state each time.  Requires code I just
committed to the GUI to function properly.

Also, add a sleep after submitting and requesting text
blobs (bad me, will have to find a better way to do this
in the future so that we don't depend on sleep and
races).

Signed-off-by: Ben Greear <greearb@candelatech.com>