cv:  Add tr398 automation test
This will launch and configure tr398 and gather the results
in the end, similar to other chamber-view tests.

This one utilizes a config file to store the calibration and radio
and attenuator config, see example-configs/tr398-ferndale-ac-cfg.txt

A new calibration file would need to be used for your own testbed.

Signed-off-by: Ben Greear <greearb@candelatech.com>