test_l3_longevity.py : On AP: No stations are currently associated. No stations are currently associated, causes an exception when reading the AP for the MAC address.
There are timing conditions when reading an AP/DUT.
The state of the DUT is read from serial, ssh, Telnet. The script expects
the AP/DUT in a specific state.  The DUT state is not guaranteed.
What may help is for the customer API to interface to the Device Under Test
So as to isolate these timing issues and do log scraping for the correct
Response.