generic-cx:  code cleanup and return error checking.
Clean up monitor logic in gen_cx profile to remove useless sleeps
and rename a variable that indicated it was msec but which was actually
seconds.

test_generic gets better error checking and also fixes bugs in how it
monitors the connection.  More work yet is needed for making the
monitor do a good pass/fail logic.

Signed-off-by: Ben Greear <greearb@candelatech.com>