regression-test:  Improve lf_dataplane_test
Use DUT that earlier test logic created, allow specifying testbed
name on the cmd line.

Add exit code for dpt, but not sure it is correct since currently
it fails trying to copy the results locally since my GUI is on localhost
and LANforge system is elsewhere.  Added TODO to allow this usage.

Signed-off-by: Ben Greear <greearb@candelatech.com>