l3_video_em.pl: max_tx now distinguishes max tx rate from max then follow rate.
max_tx rate was set to follow station tx rate as a manner of cheating by allowing the server side to act as if it was being communicated to what a station side maximum would be. This does not emulate an intermediate network.

Now script now defaults to max_tx meaning maximum server tx. By adding a plus into the number (--max_tx 30m+) the script will query the rxtx rate of the opposite station to see what it could transmit without undue overtransmission.

Signed-off-by: Jed Reynolds <jed@bitratchet.com>