create_station.py: Make key_mgmt option match GUI
Options in GUI display a different value than the one interpreted
by the server. Create a mapping internal to the script to
map from GUI display value to value that the server can interpret.

Without this patch, a user would have to click "Apply" or "Ok"
in the GUI for the desired key management to take effect.

Verified:
./create_station.py \
        --mgr           192.168.200.98 \
        --radio         1.1.wiphy1 \
        --num_stations  1 \
        --ssid          lanforge \
        --security      WPA2 \
        --key_mgmt      'FT-PSK (11r)' \
        --password      lanforge

Signed-off-by: Alex Gavin <alex.gavin@candelatech.com>