create_station.py: Make groupwise_cipher option match GUI
Similar to other options like '--key_mgmt', the GUI displays
a different value than the one interpreted by the server.

This patch follows a previously-used, backwards-compatible
mapping method, allowing the user to specify the groupwise cipher
as either the string displayed in the GUI or directly,
as was done before this patch.

Signed-off-by: Alex Gavin <alex.gavin@candelatech.com>