heatmon: More robust trinkey detection plus cleaner read logic
Improved variable management, fix hardcoded /dev/ttyACM0 path,
improve detection logic for whether or not the trinkey device is
plugged in, and switch to mapfile to bypass the header lines instead
of brute force reading 5 lines. Relies on a udev rule to setup a symlink
for the device path into /dev/TrinkeySHT41.